feat: use DragInt for nit controls (1-nit precision), widen ImGui panels

+++ src/sdr_app.h
@@ -23,5 +23,5 @@ private:
     VkBuffer quadVB_ = VK_NULL_HANDLE;
     VkDeviceMemory quadVBMem_ = VK_NULL_HANDLE;
     float sdrAlpha_ = 1.0f;
-    float paperWhite_ = 500.0f;
+    int paperWhite_ = 500;
 };
